Move PayloadBuilder to PayloadBuilder.ts
